try-catch/throw/if changed. 2 bug fix. some revision on doc
now, lessTyping can throw and catch anything
throw 'a string'
and try-catch-finally is changed to this form:
try
    ...
catch e
    ...
finally
    ...
the exception types are not required. and the caught 'e' can be any object.
if is now parsed as
a ifeq (goto nop1)
A
goto end
nop1
b ifeq (goto nop2)
B
goto end
nop2
c ...
...
D
end

fix:
fixed a bug that primitive types now can be correctly retrieved
fixed a bug that try-catch fails because of original parsing method of 'if'

error recovery for parser is in development